requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
Deprecated==1.2.13
Django==4.0.6
django-debug-toolbar==3.5.0
djangorestframework==3.13.1
Faker==13.15.1
gunicorn==20.1.0
hiredis==2.0.0
idna==3.3
packaging==21.3
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
redis==4.3.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.11
wrapt==1.14.1
